% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_incidence.R
\name{check_incidence}
\alias{check_incidence}
\title{Fit Bayesian 3-level Poisson model to evaluate incidence}
\usage{
check_incidence(data, iter = 500, warmup = 100, chains = 4, cores,
  seed = 123, nsimul)
}
\arguments{
\item{data}{Data file.}

\item{iter}{A positive integer specifying how many iterations for each chain
(including warmup). The default is 500.}

\item{warmup}{a positive integer specifying number of warmup (aka burnin)
iterations. Warmup samples should not be used for inference. The number of
warmup should not be larger than \code{iter} and the default is 100.}

\item{chains}{A positive integer specifying number of chains. Defaults to 4.}

\item{cores}{Number of cores to use when executing the chains in parallel (up
to the number of chains).}

\item{seed}{Positive integer. Used by \code{set.seed} to make results
reproducible.}

\item{nsimul}{Number of simulations.}
}
\value{
An object of class \code{stanfit}.
}
\description{
Fit a Bayesian 3-level Poisson model using Stan to evaluate true incidence,
quantify bias, and compare relative impact of selection vs. misclassification
biases.
}
\examples{
sim_list <- vector("list", 5)
set.seed(123)
sim_list <- replicate(n = 5, expr = make_data(100, 30, "saureus"), simplify = FALSE)
check_incidence(sim_list,
                iter = 200,
                warmup = 25,
                chains = 2,
                cores = 2,
                seed = 123,
                nsimul = 5)
}
\author{
Denis Haine
}

