% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misclass.R
\name{coef.misclass}
\alias{coef.misclass}
\title{Extract model coefficients}
\usage{
\method{coef}{misclass}(object, parm, method = "Median", ...)
}
\arguments{
\item{object}{a misclass object}

\item{parm}{parameter for which coefficients are required. Defaults to all}

\item{method}{Method for estimating the center of the posterior distribution.
Defaults to "Median", the alternative is "Mean"}

\item{...}{ignored}
}
\description{
Extract model coefficients from a misclass object.
}
