% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misclass.R
\name{confint.misclass}
\alias{confint.misclass}
\title{Credible intervals}
\usage{
\method{confint}{misclass}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{a misclass object}

\item{parm}{vector for names of parameters for which credibile intervals are
requested. Defaults to all.}

\item{level}{the credible interval required}

\item{...}{ignored}
}
\description{
Credible intervals for a misclass object.
}
\details{
Unlike the generic function \code{confint}, this function produces
credible intervals, not confidence intervals. The intervals are estimated
using the \code{quantile} of the mcmc samples.
}
