% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misclass.R
\name{extend.misclass}
\alias{extend.misclass}
\title{Extend the mcmc chain of a misclass object}
\usage{
extend.misclass(
  object,
  sample = object$jags.options$sample,
  warmup = 0,
  adapt = 0
)
}
\arguments{
\item{object}{a \code{misclass} object}

\item{sample}{Number of samples by which the mcmc chain(s) should be extended,
defaults to same as original object (i.e. the total of samples is doubled).}

\item{warmup}{Number of additional warmup iterations for the extension of the
chain, defaults to 0.}

\item{adapt}{Number of additional adaptation iterations of the mcmc sampler,
defaults to 0.}
}
\description{
When the \code{misclass} object contains too few samples, this function can
extend the MCMC chain. By default, it doubles the total amount of samples
used for inference, but does not perform another warmup or adaptation.
}
\details{
see \code{extend.jags} from the {runjags} package for details.
}
