% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{make.model}
\alias{make.model}
\title{Make a misclassification model for JAGS}
\usage{
make.model(
  measurement = list(stratified.differential = FALSE, additive.differential = FALSE,
    covariate = TRUE, intercept.random = FALSE),
  exposure = list(covariate = TRUE, intercept.random = FALSE),
  outcome = list(covariate = TRUE, intercept.random = FALSE, effects.random = FALSE),
  priors = list(coef.precision = 0.1, heterogeneity.precision = 5)
)
}
\arguments{
\item{measurement}{List for measurement submodel, containing:
\code{stratified.differential} to define a stratified differential
misclassification model (i.e. the measurement submodel is stratified by
values of y);
\code{additive.differential} to define a differential misclassification
model with additive terms for y in the measurement submodel (i.e. linear
on the log-odds scale);
\code{covariate} to allow for covariate effects (for z);
and \code{intercept.random} to allow for random intercepts in the
measurement submodel. All may be abbreviated to the first letter.}

\item{exposure}{List for exposure submodel, containing:
\code{covariate} to allow for covariate effects (for z) in the exposure
submodel;
and \code{intercept.random} to allow for random intercepts in the exposure
submodel. All may be abbreviated to the first letter.}

\item{outcome}{List for outcome submodel, containing:
\code{covariate} to allow for covariate effects (for z) in the outcome
submodel;
\code{intercept.random} to allow for random intercepts in the outcome
submodel;
and \code{effects.random} to allow for random effects (random slopes) in
the outcome submodel. All may be abbreviated to the first letter.}

\item{priors}{List for priors, containing \code{coef.precision}: the value
for the precision of the priors for the coefficients, and
\code{heterogeneity.precision}: the value for the precision of the priors
for the heterogeneity parameters.}
}
\value{
a JAGS model, as a character string that \code{run.jags} from the
\code{runjags} package can read.
}
\description{
Make a misclassification model that runjags/JAGS can read. The model can have
covariate effects and random intercepts for all submodels and also random
slopes for the outcome submodel. To view the model, save it to a file and
open it (see examples).
}
\details{
\code{stratified.differential} Allows for completely
different submodels for different values of y (and x), and thereby
overrides \code{additive.differential}.

The output can be saved as .R or .txt file, making
it easier to read, and allowing it to be run with rjags as well. See the
examples.
}
\examples{
# Default: Nondifferential error, covariate effects but no random effects.
m <- make.model() 

# Random intercepts. Note, not the whole list needs to be entered, only the 
# arguments that are to be changed. 
m <- make.model(measurement = list(intercept.random = TRUE), 
exposure = list(intercept.random = TRUE), 
outcome = list(intercept.random = TRUE))

# Random effects. Note that options of the lists can be abbreviated to the 
# first letter.
m <- make.model(measurement = list(i = TRUE), 
exposure = list(i = TRUE), 
outcome = list(i = TRUE, e = TRUE))
               
# Differential misclassification
m <- make.model(measurement = list(s = TRUE))
m <- make.model(measurement = list(a = TRUE))

\dontrun{
write(m, file = "jags_script.R")
file.edit("jags_script.R")}
}
