% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitors.R
\name{make.monitors}
\alias{make.monitors}
\title{Make parameters for monitoring}
\usage{
make.monitors(
  measurement = list(stratified.differential = FALSE, additive.differential = FALSE,
    covariate = TRUE, intercept.random = FALSE),
  exposure = list(covariate = TRUE, intercept.random = FALSE),
  outcome = list(intercept.random = FALSE, slope.random = FALSE, covariate = TRUE),
  monitor.re = FALSE
)
}
\arguments{
\item{measurement}{List for measurement submodel, containing:
\code{stratified.differential} to define a stratified differential
misclassification model (i.e. the measurement submodel is stratified by
values of y);
\code{additive.differential} to define a differential misclassification
model with additive terms for y in the measurement submodel (i.e. linear
on the log-odds scale);
\code{covariate} to allow for covariate effects (for z);
and \code{intercept.random} to allow for random intercepts in the
measurement submodel. All may be abbreviated to the first letter.}

\item{exposure}{List for exposure submodel, containing:
\code{covariate} to allow for covariate effects (for z) in the exposure
submodel;
and \code{intercept.random} to allow for random intercepts in the exposure
submodel. All may be abbreviated to the first letter.}

\item{outcome}{List for outcome submodel, containing:
\code{covariate} to allow for covariate effects (for z) in the outcome
submodel;
\code{intercept.random} to allow for random intercepts in the outcome
submodel;
and \code{effects.random} to allow for random effects (random slopes) in
the outcome submodel. All may be abbreviated to the first letter.}

\item{monitor.re}{Should the estimated random effects per clusters be
monitored? Defaults to FALSE, as this produces a lot of output when many
clusters are present.}
}
\value{
Returns a List of parameter names
}
\description{
Make parameters for monitoring the JAGS object in misclass.
}
\examples{
# Nondifferential model, with no random effects or random intercepts.
m <- make.monitors()

# Differential misclassification model with random intercepts, random effects,
# fixed covariate effects 
m <- make.monitors(measurement = list(stratified.differential = TRUE,
                                      additive.differential   = FALSE,
                                      covariate               = TRUE,
                                      intercept.random        = TRUE),
                   exposure    = list(covariate               = TRUE,
                                      intercept.random        = TRUE),
                   outcome     = list(intercept.random        = TRUE,
                                      slope.random            = TRUE,
                                      covariate               = TRUE))
# Random intercepts (abbreviated to i), and covariate effects (which default 
# to TRUE) for all three submodel.
m <- make.monitors(measurement = list(i = TRUE),
                   exposure    = list(i = TRUE),
                   outcome     = list(i = TRUE))

}
