% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeData.R
\name{makeData}
\alias{makeData}
\title{Create simulated datasets.}
\usage{
makeData(n_herd, n_cow, E_hPr = 0.5, E_cPr = 0.5, E_qPr = 0.5,
  sigma_sqhPr = 0.14, sigma_sqcPr = 2.25, b0_Pr = -6.7, OR_hPr = 3,
  OR_cPr = 3, OR_qPr = 3, sigma_sqhI = 0.838, sigma_sqcI = 2.926,
  b0_I = -8.3, OR_hI = 3, OR_cI = 3, OR_qI = 3, sigma_sqhEl = 0.15,
  sigma_sqcEl = 2.246, b0_El = -0.6, se_parms = c(0.9, 0.85),
  sp_parms = 1, se_series = -0.1, sp_series = 0, se_parall = 0.1,
  sp_parall = 0, se_tri = 0, sp_tri = 0)
}
\arguments{
\item{n_herd}{Number of herds.}

\item{n_cow}{Number of cows per herd.}

\item{E_hPr}{Exposure distribution (0 to 1) of the binary herd-level (h) predictor.}

\item{E_cPr}{Exposure distribution (0 to 1) of the binary cow-level (c) predictor.}

\item{E_qPr}{Exposure distribution (0 to 1) of the binary quarter-level (q) predictor.}

\item{sigma_sqhPr}{Herd-level variance (sigma_sq) for prevalence of intra-mammary infection (IMI).}

\item{sigma_sqcPr}{Cow-level variance for prevalence of IMI.}

\item{b0_Pr}{Intercept for IMI prevalence; aiming at a prevalence of 2.5%.}

\item{OR_hPr}{OR of association between herd-level variable and IMI prevalence.}

\item{OR_cPr}{OR of association between cow-level variable and IMI prevalence.}

\item{OR_qPr}{OR of association between observation-level variable and IMI prevalence.}

\item{sigma_sqhI}{Herd-level variance for incidence of IMI.}

\item{sigma_sqcI}{Cow-level variance for incidence of IMI.}

\item{b0_I}{Intercept for IMI incidence.}

\item{OR_hI}{OR of association between herd-level variable and IMI incidence.}

\item{OR_cI}{OR of association between cow-level variable and IMI incidence.}

\item{OR_qI}{OR of association between observation-level variable and IMI incidence.}

\item{sigma_sqhEl}{Herd-level variance for elimination of IMI.}

\item{sigma_sqcEl}{Cow-level variance for elimination of IMI.}

\item{b0_El}{Intercept for IMI persistency.}

\item{se_parms}{Vector for mode and x to determine shape parameters of Beta distribution of sensitivity (Se).}

\item{sp_parms}{Vector for mode and x to determine shape parameters of Beta distribution of specificity (Sp).}

\item{se_series}{Se improvement or loss associated with sampling strategy: duplicate series.}

\item{sp_series}{Sp improvement or loss associated with sampling strategy: duplicate series.}

\item{se_parall}{Se improvement or loss associated with sampling strategy: duplicate parallel.}

\item{sp_parall}{Sp improvement or loss associated with sampling strategy:duplicate parallel.}

\item{se_tri}{Se improvement or loss associated with sampling strategy: triplicate (2 out of 3).}

\item{sp_tri}{Sp improvement or loss associated with sampling strategy: triplicate (2 out of 3).}
}
\value{
A data frame with variables:
\item{herd}{Herd id.}
\item{cow}{Cow id.}
\item{quarter}{Quarter id.}
\item{S1}{First milk sample true status.}
\item{S2}{Second milk sample true status.}
\item{E_h}{Herd-level exposure.}
\item{E_c}{Cow-level exposure.}
\item{E_q}{Quarter-level exposure.}
\item{S1i}{Misclassified first milk sample.}
\item{S2i}{Misclassified second milk sample.}
\item{S1_series}{Misclassified first milk sample based on duplicate series sampling strategy.}
\item{S2_series}{Misclassified second milk sample based on duplicate series sampling strategy.}
\item{S1_parall}{Misclassified first milk sample based on duplicate parallel sampling strategy.}
\item{S2_parall}{Misclassified second milk sample based on duplicate parallel sampling strategy.}
\item{S1_tri}{Misclassified first milk sample based on triplicate sampling strategy.}
\item{S2_tri}{Misclassified second milk sample based on triplicate sampling strategy.}
}
\description{
Create datasets from a hypothetical cohort study. Since correlation structures
are often found in udder health studies, the generated datasets are be deemed
to have occurred from the collection of two milk samples collected 1 month
apart from each quarters of a random sample of 30 cows per herd, from 100
dairy herds. The first milk sample (S1) is used to identify quarters at risk of
intramammary infection (IMI) at the beginning of the cohort, while the second
(S2) is used to identify the outcome (acquisition of a new IMI). Three
hypothetical exposures E_q, E_c, E_h (quarter, cow, and herd level) with known
strength of association (OR~3.0) are generated. As it is often the case
(Dufour et al., 2011, 2012), exposures are equally associated with odds of a
prevalent IMI on first milk sample as with odds of IMI acquisition on second
sample. Exposures are randomly associated with odds of eliminating an existing
IMI (OR=1.0). By default, dataset for S. aureus is created.
}
\examples{
# Initiate a list to store the n data frames
sim_list <- vector("list", 5)
# Do not forget to set seed for replication
set.seed(123)
sim_list <- replicate(n = 5, expr = makeData(100, 30), simplify = FALSE)
# Or with a progress bar
require(pbapply)
sim_list <- pbreplicate(n = 5, expr = makeData(100, 30), simplify = FALSE)
}
\references{
Dufour, S., Dohoo, I.R., Barkema, H.W., DesCôteaux, L., DeVries, T.J., Reyher, K.K., Roy, J.-P., Scholl, D.T., 2012 \emph{Epidemiology of coagulase-negative staphylococci intramammary infection in dairy cattle and the effect of bacteriological culture misclassification}. Journal of Dairy Science 95(6):3110-3124.
}

