\name{misclass}
\alias{misclass}
\alias{misclass-package}
\docType{package}
\title{Sampling strategies to control misclassification bias in longitudinal udder health studies}
\description{Code to run Bayesian evaluation  of misclassification bias.}
\details{
  \tabular{ll}{
    Package: \tab misclass\cr
    Type: \tab Package\cr
    Version: \tab 0.0.0.9000\cr
    Date: \tab 2016-05-19\cr
    License: \tab GPL-2\cr
  }
}
\author{
  Denis Haine \email{denis.haine@gmail.com}
}
