% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misclass.R
\name{misclass}
\alias{misclass}
\title{Misclassification model for clustered binary data}
\usage{
misclass(
  x,
  s,
  y,
  z,
  j,
  n.chains = 2,
  adapt = 1000,
  warmup = 5000,
  sample = 10000,
  thin = 1,
  psrf.max = 1.05,
  monitor.re = FALSE,
  monitor = NULL,
  inits = NULL,
  inits.options = NULL,
  measurement = NULL,
  exposure = NULL,
  outcome = NULL,
  priors = NULL,
  model = NULL,
  ...
)
}
\arguments{
\item{x}{Gold standard measurement of the binary exposure, possibly missing.}

\item{s}{Surrogate measurement of the binary exposure. Potentially
misclassified.}

\item{y}{Binary outcome y. Possibly missing.}

\item{z}{Matrix or a single vector of covariates. If a matrix, each column
should be one covariate, each row an observation/participant/patient.}

\item{j}{Indicator variable for the cluster (country, center, data set) the
observation belongs to. If missing, all samples are assumed to belong to
the same cluster.}

\item{n.chains}{Number of MCMC chains}

\item{adapt}{Number of iterations for the sampler to adapt before running the
warmup and final analysis.}

\item{warmup}{Number of iterations for the sampler to find the posterior,
before running the final analysis (also referred to as burnin).}

\item{sample}{Number of iterations for final analysis.}

\item{thin}{thinning interval. Values > 1 can be used to discard samples to
reduce autocorrelation.}

\item{psrf.max}{maximum value for the psrf and mpsrf, to automatically
to flag non-convergence quantified by \code{gelman.diag} from the
\code{coda} package..}

\item{monitor.re}{Should the estimated random effects per cluster be
monitored?}

\item{monitor}{Optional vector for setting monitors. Defaults to all fixed
effects, and standard deviations for random effects/intercepts. Include
include all random by setting \code{monitor.re} to \code{TRUE}.}

\item{inits}{Optional initial values. If left \code{NULL} separate initial
values are generated for each chain, using \code{inits.options}}

\item{inits.options}{List of: \code{coef.init.limit} For coefficients,
initial values are generated between coef.init.limit and coef.init.limit;
\code{tau2.init.min} Minimum for initial values of tau2 heterogeneity
parameters; and \code{tau2.init.max} Maximum for initial values of tau2
heterogeneity}

\item{measurement}{List for measurement submodel, containing:
\code{stratified.differential} to define a stratified differential
misclassification model (i.e. the measurement submodel is stratified by
values of y);
\code{additive.differential} to define a differential misclassification
model with additive terms for y in the measurement submodel (i.e. linear
on the log-odds scale);
\code{covariate} to allow for covariate effects (for z);
and \code{intercept.random} to allow for random intercepts in the
measurement submodel. All may be abbreviated to the first letter.}

\item{exposure}{List for exposure submodel, containing:
\code{covariate} to allow for covariate effects (for z) in the exposure
submodel;
and \code{intercept.random} to allow for random intercepts in the exposure
submodel. All may be abbreviated to the first letter.}

\item{outcome}{List for outcome submodel, containing:
\code{covariate} to allow for covariate effects (for z) in the outcome
submodel;
\code{intercept.random} to allow for random intercepts in the outcome
submodel;
and \code{effects.random} to allow for random effects (random slopes) in
the outcome submodel. All may be abbreviated to the first letter.}

\item{priors}{List for priors, containing \code{coef.precision}: the value
for the precision of the priors for the coefficients, and
\code{heterogeneity.precision}: the value for the precision of the priors
for the heterogeneity parameters.}

\item{model}{Optional JAGS model. Overrides \code{measurement},
\code{exposure}, \code{outcome}, and \code{priors},}

\item{...}{Optional parameters for \code{run.jags} from the \code{runjags}
package.}
}
\description{
Misclassification model for clustered individual participant data with binary
outcomes (endpoints), continuous or binary covariates and a binary exposure
(or intervention) that is possibly misclassified in some studies.
}
\details{
This function (and the rest of the package) is intended to make
using a misclassification model easier. Therefore, it is not entirely
comprehensive.  However, the package is designed to be flexible, so that
you can use partial outputs (the JAGS model, monitors and inits), if you
would like to tweak the JAGS model.

For the coefficients and random intercepts/effects normal priors are used.
For the variances of the heterogeneity across clusters, half-normal priors
are used.
}
