% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runProgram.R
\name{runProgram}
\alias{runProgram}
\title{Fit Bayesian 3-level logistic model to evaluate misclassification.}
\usage{
runProgram(data, iter = 1000, warmup = 100, chains = 4, cores,
  seed = 123)
}
\arguments{
\item{data}{Data file.}

\item{iter}{A positive integer specifying how many iterations for each chain
(including warmup). The default is 1000.}

\item{warmup}{A positive integer specifying number of warmup (aka burnin)
iterations. Warmup samples should not be used for inference. The number of
warmup should not be larger than iter and the default is 100.}

\item{chains}{A positive integer specifying number of chains. Defaults to 4.}

\item{cores}{Number of cores to use when executing the chains in parallel (up
to the number of chains).}

\item{seed}{Positive integer. Used by \code{set.seed} to make results
reproducible.}
}
\value{
An object of class \code{stanfit}.
}
\description{
Fit a Bayesian 3-level logistic model using Stan to evaluate true association,
quantify bias, and compare relative impact of selection vs. misclassification
biases.
}
\examples{
runProgram(sim_list[[1]],
           iter = 200,
           warmup = 25,
           chains = 4,
           cores = 4,
           seed = 123)
}
\author{
Denis Haine
}

