% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sample.misclass}
\alias{sample.misclass}
\title{Sample clustered binary data with misclassified exposure}
\usage{
sample.misclass(
  n = 50,
  J = 1,
  beta.0.0 = NULL,
  tau.beta.0.j = 0,
  beta.x.0 = log(2),
  tau.beta.x.j = 0,
  beta.z.0 = log(2),
  tau.z.mu = 1/2,
  cov = 1/4,
  gamma.0.0 = 0,
  tau.gamma.0.j = 0,
  gamma.z.0 = 1/2,
  lambda.0.0 = 3,
  tau.lambda.0.j = 1,
  lambda.z.0 = -2,
  phi.0.0 = -3,
  tau.phi.0.j = 1,
  phi.z.0 = 2,
  colnames = NULL,
  center.z = TRUE
)
}
\arguments{
\item{n}{sample size per cluster}

\item{J}{number of clusters}

\item{beta.0.0}{intercept in the outcome model. if NULL, the incidence is
fixed so that mean(y) converges to 1/2.}

\item{tau.beta.0.j}{variance of the random intercept in the outcome model}

\item{beta.x.0}{coefficient for x in the outcome model}

\item{tau.beta.x.j}{variance of random effect for x in the outcome model}

\item{beta.z.0}{log odds ratios for covariates.}

\item{tau.z.mu}{SD of the mean of the z covariates across studies}

\item{cov}{Covariance for continuous covariates}

\item{gamma.0.0}{intercept in the exposure model}

\item{tau.gamma.0.j}{variance of the random intercept in the exposure model}

\item{gamma.z.0}{coefficients for z in the exposure model}

\item{lambda.0.0}{intercept in the measurement model, if x = 1}

\item{tau.lambda.0.j}{variance of the random intercept in the measurement
model, if x = 1}

\item{lambda.z.0}{coefficients in the measurement model, if x = 1}

\item{phi.0.0}{intercept in the measurement model, if x = 0}

\item{tau.phi.0.j}{variance of the random intercept in the measurement model,
if x = 0}

\item{phi.z.0}{coefficients in the measurement model, if x = 0}

\item{colnames}{optional column names}

\item{center.z}{should covariates be centered? (post hoc)}
}
\value{
A data.frame, including the mismeasured and correctly measured
variables.
}
\description{
Sample clustered binary data with misclassified exposure, correctly
classified exposure, covariates and outcome. Random intercepts for outcome,
measurementand exposure model. Random effects for outcome model.
}
\examples{
d <- sample.misclass(gamma.z.0 = c(1/2, 1/2),
lambda.z.0 = c(-1, -1),
phi.z.0 = c(1, 1),
beta.z.0 = c(1, 1))
}
