% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_incidence.R
\name{sample_incidence}
\alias{sample_incidence}
\title{Fit Bayesian 3-level Poisson model to evaluate sampling strategy for
misclassification in incidence evaluation.}
\usage{
sample_incidence(data, iter = 500, warmup = 100, chains = 4, cores,
  seed = 123, nsimul)
}
\arguments{
\item{data}{Data file.}

\item{iter}{A positive integer specifying how many iterations for each chain
(including warmup). Default is 500.}

\item{warmup}{A positive integer specifying number of warmup (aka burnin)
iterations. Warmup samples should not be used for inference. The number of
warmup should not be larger than \code{iter} and the default is 100.}

\item{chains}{A positive integer specifying number of chains. Defaults to 4.}

\item{cores}{Number of cores to use when executing the chains in parallel (up
to the number of chains).}

\item{seed}{Positive integer. Used by \code{set.seed} to make results
reproducible.}

\item{nsimul}{Number of simulations.}
}
\value{
An object of class \code{stanfit}.
}
\description{
Fit a Bayesian 3-level Poisson model using Stan to evaluate effect of various
sampling strategies on biases when looking at incidences.
}
\examples{
sample_incidence(sim_list[[1]],
                 iter = 200,
                 warmup = 25,
                 chains = 4,
                 cores = 4,
                 seed = 123)
}
\author{
Denis Haine
}

