% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_stgy.R
\name{sample_stgy}
\alias{sample_stgy}
\title{Fit Bayesian 3-level logistic model to evaluate sampling strategies to correct
for presence of bias of measure of association.}
\usage{
sample_stgy(data, iter = 500, warmup = 100, chains = 4, cores,
  seed = 123, nsimul)
}
\arguments{
\item{data}{Data file.}

\item{iter}{A positive integer specifying how many iterations for each chain
(including warmup). Default is 500.}

\item{warmup}{A positive integer specifying number of warmup (aka burnin)
iterations. Warmup samples should not be used for inference. The number of
warmup should not be larger than iter and the default is 100.}

\item{chains}{A positive integer specifying number of chains. Defaults to 4.}

\item{cores}{Number of cores to use when executing the chains in parallel (up
to the number of chains).}

\item{seed}{Positive integer. Used by \code{set.seed} to make results
reproducible.}

\item{nsimul}{Number of simulations.}
}
\value{
An object of class \code{stanfit}.
}
\description{
Fit a Bayesian 3-level logistic model using Stan to evaluate effect of various
sampling strategies on biases of measure of association.
}
\examples{
sim_list <- vector("list", 1)
set.seed(123)
sim_list <- replicate(n = 1, expr = make_data(100, 30, "saureus"), simplify = FALSE)
\dontrun{sample_stgy(sim_list,
                     iter = 200,
                     warmup = 25,
                     chains = 1,
                     cores = 1,
                     seed = 123,
                     nsimul = 1)}
}
\author{
Denis Haine
}

